requirements.txt
numpy
keras
tensorflow-cpu==2.5.0
pandas
matplotlib
pillow
flask
seaborn
gunicorn
